requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
celery>=5.3.6
redis>=5.0.1
python-multipart>=0.0.9
google-api-python-client>=2.116.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
face_recognition>=1.3.0
numpy>=1.26.0
httpx>=0.26.0
python-socketio>=5.11.0
requests>=2.31.0
dlib>=19.24.2
pydantic-settings>=2.0.0
